Fix for groups being deleted because they appeared to be empty
Addressables configuration data was being read while addressable configuration files were being imported. This leads to groups appearing to be empty when they're not - just because the data hasn't been read in yet.
This commit avoids running the deletion part of the post processor if one of these files is being imported.